refactor(main-header): unify header menu as cross-platform heroui Menu
Replace the platform-split menu (Android MenuView / iOS Stack.Toolbar.Menu) in MainHeader with a single heroui Menu popover, matching the MessageHeader pattern. Pin state maps to isSelected plus title toggle; back/new buttons preserved.
